feature: Enumerator, fix: Repository Error, feature: Value Dto/Validator

+++ backend/source/php/repository/EnumeratorRepository.php
@@ -5,42 +5,51 @@ namespace mate\repository;
 use mate\abstract\clazz\Repository;
 use mate\abstract\trait\SelectByName;
 use mate\error\WPErrorBuilder;
-use mate\model\UnitModel;
+use mate\model\EnumeratorModel;
 use PDO;
 use Throwable;
 
-class UnitRepository extends Repository
+class EnumeratorRepository extends Repository
 {
     use SelectByName;
 
-    protected string $table = "mate_unit";
-    protected string $model = UnitModel::class;
+    protected string $table = "mate_enumerator";
+    protected string $model = EnumeratorModel::class;
 
-    private readonly UnitValueRepository $valueRepository;
+    private readonly EnumeratorValueRepository $valueRepository;
 
     public function __construct()
     {
         parent::__construct();
-        $this->valueRepository = UnitValueRepository::inject();
+        $this->valueRepository = EnumeratorValueRepository::inject();
     }
 
     public function insert($model): ?object
     {
         $this->db->transaction();
         $q = <<<SQL
-        INSERT INTO {$this->table} (`name`, `description`)
-        VALUES (:name, :description)
+        INSERT INTO {$this->table} (
+            `name`,
+            `description`,
+            `typeId`
+        )
+        VALUES (
+            :name,
+            :description,
+            :typeId
+        )
         SQL;
 
         try {
             $s = $this->db->prepare($q);
             $s->bindValue(":name", $model->name, PDO::PARAM_STR);
             $s->bindValue(":description", $model->description, PDO::PARAM_STR);
+            $s->bindValue(":typeId", $model->typeId, PDO::PARAM_INT);
             $s->execute();
             $model->id = $this->db->lastInsertId();
 
             foreach ($model->valueList as $value) {
-                $value->unitId = $model->id;
+                $value->enumeratorId = $model->id;
                 $this->valueRepository->insert($value);
             }
 
@@ -56,9 +65,14 @@ class UnitRepository extends Repository
     {
         $this->db->transaction();
         $previousModel = $this->selectById($model->id);
+        $updateType = $previousModel->typeId !== $model->typeId;
+
         $q = <<<SQL
         UPDATE {$this->table}
-        SET `name` = :name, `description` = :description
+        SET
+        `name` = :name,
+        `description` = :description,
+        `typeId` = :typeId
         WHERE `id` = :id
         SQL;
 
@@ -69,23 +83,31 @@ class UnitRepository extends Repository
             $s = $this->db->prepare($q);
             $s->bindValue(":name", $model->name, PDO::PARAM_STR);
             $s->bindValue(":description", $model->description, PDO::PARAM_STR);
+            $s->bindValue(":typeId", $model->typeId, PDO::PARAM_INT);
             $s->bindValue(":id", $model->id, PDO::PARAM_INT);
             $s->execute();
 
+            if ($updateType) {
+                $this->valueRepository->deleteByEnumeratorId($model->id);
+            }
+
             foreach ($model->valueList as $value) {
-                $value->unitId = $model->id;
+                $value->enumeratorId = $model->id;
                 if ($value->id !== null) {
-                    $this->valueRepository->update($value);
+                    if ($updateType === false) {
+                        $this->valueRepository->update($value);
+                    }
                 } else {
                     $this->valueRepository->insert($value);
                 }
             }
 
-
-            array_map(
-                [$this->valueRepository, "deleteById"],
-                array_filter($previousValueIdList, fn ($id) => !in_array($id, $newValueIdList))
-            );
+            if ($updateType === false) {
+                array_map(
+                    [$this->valueRepository, "deleteById"],
+                    array_filter($previousValueIdList, fn($id) => !in_array($id, $newValueIdList))
+                );
+            }
 
             $this->db->commit();
             return $this->selectById($model->id, false);
@@ -96,12 +118,12 @@ class UnitRepository extends Repository
         }
     }
 
-    public function selectById(int $id, bool $cache = true): ?UnitModel
+    public function selectById(int $id, bool $cache = true): ?object
     {
         $m = parent::selectById($id, $cache);
 
         if ($m !== null) {
-            $m->valueList = $this->valueRepository->selectByUnitId($m->id);
+            $m->valueList = $this->valueRepository->selectByEnumeratorId($m->id);
         }
 
         return $m;
